price att

+++ api/models.py
@@ -8,7 +8,7 @@ class Produto(models.Model):
     nome = models.CharField(max_length=255)
     descricao = models.TextField()
     quantidades = models.IntegerField()
-    preco = models.FloatField()
+    preco = models.DecimalField(max_digits=10, decimal_places=2)
     categoria = models.CharField(max_length=255)
 
     def __str__(self):
